+++ profile/src/main/java/com/bank/profile/repository/ProfileRepository.java
@@ -4,7 +4,9 @@ import com.bank.profile.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile, Long> {
+public interface ProfileRepository extends JpaRepository<Profile, BigInteger> {
     Profile findProfileByInnOrPhoneNumber(Long inn, Long n);
 }
